clb-385 fixed BigInteger to Long cast error when a load balancer exists in stingray but not in our database.

+++ persistence/src/main/java/org/openstack/atlas/service/domain/usage/repository/LoadBalancerMergedHostUsageRepository.java
@@ -97,7 +97,7 @@ public class LoadBalancerMergedHostUsageRepository {
         pollTimeCal.setTimeInMillis(pollTimeMillis);
 
         LoadBalancerMergedHostUsage usageItem = new LoadBalancerMergedHostUsage();
-        usageItem.setId((Long) row[0]);
+        usageItem.setId(((BigInteger) row[0]).longValue());
         usageItem.setLoadbalancerId((Integer) row[1]);
         usageItem.setConcurrentConnections(((BigInteger) row[2]).longValue());
         usageItem.setIncomingTransfer(((BigInteger) row[3]).longValue());
